mm: Drop Google's picks and put do_wp_page back in line with stable
This causes TEEGRIS to crash out and die. The user-visible result is
that the lockscreen pin is claimed to be incorrect, making the tablet
unusable. The device may be unlocked once upon boot, if done quickly
enough, but not after relocking or sitting long enough after boot.

This is a squash of:
* Revert "UPSTREAM: mm: reuse only-pte-mapped KSM page in do_wp_page()"
  This reverts commit 2b438ad3d5bad3d42b772d79f1aed4e143fe4c75.
* Revert "UPSTREAM: mm/ksm: Remove reuse_ksm_page()"
  This reverts commit e374fe2602945f563e12ba9e3276f196523a4802.
* Revert "BACKPORT: mm: do_wp_page() simplification"
  This reverts commit 32c2bf987549e56f11784b194a33b091fa3bc482.
* Revert "UPSTREAM: mm: fix misplaced unlock_page in do_wp_page()"
  This reverts commit ecff1109d781e1ef07f5b63658ce9429a0948f07.

Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Change-Id: I83980aea81a0b667068acdf6ffc7d698ff1d51bf
Signed-off-by: Rve27 <rve27github@gmail.com>